Device Первая версия рефакторинга каталога Data. Данные пока не выводит. Хоть не виснет - и то хлеб!

+++ sources/Device/src/Osci/Measurements/Measurements.cpp
@@ -1391,7 +1391,7 @@ void Osci::Measurements::SetData(bool /*needSmoothing*/)
 
     if (TBASE_DS >= TBase::MIN_P2P)           // ≈сли находимс€ в поточечном режме, то нужно брать последние считанные точки дл€ проведени€ измерений
     {
-        for (int i = NUM_BYTES_DS - 1; i >= 0; --i)
+        for (int i = (int)(NUM_BYTES_DS - 1); i >= 0; --i)
         {
             if (IN_A[i] != NONE)      // ≈сли это значение считано
             {
